Fixed a bug that led to crashes during initial sync due to incorrect calculation of deletion/insertion/modification positions for the table. Fix sync error "Index 332 out of bounds for length 332" #89 Fixes the issue where if item were deleted remotely it was not removed from the table during partial sync. Fixes the issue where if item were added remotely it was always added to the end of the table, instead of being inserted in it's correct position.
Since I now completely understand the All Items Screen business logic I decided to do a major refactoring related to generation of rows for this screen.  This led to a number of bugfixes listed above.

AllItemsProcessor class and is now responsible for all the work related to querying and listening to the database, processing the results and posting generated ItemCellModels to the ViewModel. Hence the ViewModel is much cleaner now.

ReadItemsDbRequest now doesn't return the actual result list, but instead an empty collection that is generated by Realm's findAsync method that is then being observed by AllItemsProcessor. It allows for much more straightforward and less duplicated code related to generation of ItemCellModels.