Missing page scrubber in PDF reader. Closes #113 Fixes “Darker gray background around PDF in dark mode”. Closes #59
Two Important nuances regarding PSPDFKIT:
Nuance #1.  Adjusting design of PSPDFKIT:
There are 2 ways to change colors of components in the PSPDFKIT - either dynamically at runtime via calls like “pdfThumbnailBar.thumbnailBorderColor” or statically via a “style” property in our app’s style library.
As I’ve discovered for some components color cannot be changed at all programmatically, even though the corresponding API is present (like in case of thumbnailBorderColor). In order not to waste time in the future I would highly advise first trying to change the UI of PSPDFKIT via styles, which seem consistent.

Nuance #2. Scrollbars.
As it turns out the “scrollbarsEnabled(true)” doesn’t really do much when passed via Configuration object to the PdfFragment. It probably works fine when dealing with PdfActivity, but that’s not our case.
Changing properties of parent view to account for the possibility of having a scrollable view child inside of it didn’t help. So the problem is not that the scrollbars are present, but slightly out of frame, nor is it because they are of transparent color.
As it turns out, we have to specifically declare PSPDFKIT style in our styles library and either completely setup all properties for the scrollbars ourselves, or point out to the PSPDFKIT’s own parent style. You would expect the PSPDFKIT kit to use some sensible defaults for the case where no style is specified, but for scrollbars this wasn’t the case.

Upping versionCode 49