Major rework of Tree Building logic used on CollectionScreen and CollectionPickerScreen. Solves problems: 1. Reduces number of ANRs due to a better separation of work between worker and UI threads 2. Fixes a “CollectionTree.replaceNodes java.lang.IndexOutOfBoundsException”. The root cause turned out to be a race condition. 3. Substantially reduces memory footprint and CPU utilization when building a tree during an ongoing sync process.
Previous approach completely rebuilt entire Collection Tree on every single DB change. Initially it was thought that there is no way around that because after user just logged in and we start the sync process we only get small portions of the collection rows at a time. The order we receive new rows is random and during a given sync when we try creating tree nodes from those rows some of the nodes cannot be put into a tree immediately as they might be missing one or more of their parents and grandparents (no path to root).

New approach:
But then it was realized that we can simply be putting all of the nodes with no path to root into a separate list of temporary orphans and every time new sync update arrives some of those orphans might become insertable into the tree while becoming reachable all the way from root.
Thus current approach no longer needs to rebuild entire tree on every DB update. The way this is achieved is that we listen to DB changes with Realm’s OrderedCollectionChangeSet listener which gives us deleted, modified and inserted RCollection items since the last DB update. Algorithm was developed that maps those changes onto the existing tree and correspondingly removes, modifies and adds nodes to it/from it on the fly. The new nodes that cannot be added to the tree are the ones with no path to the root and are temporarily stored in nodesWithNoPathToParent list, they are added gradually as we continue to receive DB updates and all of them will be added by the end.

Other improvements:
Children of the given node now live inside of TreeSet which automatically sorts them upon insertion modification or deletion of nodes which is much more efficient then having to walk entire tree and sorting children manually on every change.

In case of large already synced collections one major change is chunking of data send to UI to 200 operations upon the tree at a time. In most cases that just means that after 200 insertions we post current state of the tree from our data layer to UI for display so that user can immediately interact with some of the collections.

Data structures used by UI are now based on ImmutableList with a hashable data classes which allows Compose to properly calculate exactly which portion of the tree needs to be redrawn on each change reducing full tree redraws.

All of this substantially improves memory and CPU footprint especially on very large collections as it hopefully completely eliminates unnecessary steps. Especially reduces number of per item DB calls for itemCount and collection name retrievals.

Upping versionCode to 166